perbaikan rt,rw,dusun,data warga di admin

+++ app/Http/Controllers/KategoriketerampilanController.php
@@ -28,8 +28,18 @@ class KategoriKeterampilanController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'nama_kategori' => 'required|max:255',
-        ]);
+    'nama_kategori' =>
+        'required|max:255|unique:kategori_keterampilans,nama_kategori',
+],[
+    'nama_kategori.required' =>
+        'Nama kategori wajib diisi.',
+
+    'nama_kategori.unique' =>
+        'Kategori sudah ada.',
+
+    'nama_kategori.max' =>
+        'Nama kategori maksimal 255 karakter.',
+]);
 
         KategoriKeterampilan::create($request->only('nama_kategori'));
 
@@ -50,8 +60,19 @@ class KategoriKeterampilanController extends Controller
     public function update(Request $request, KategoriKeterampilan $kategori_keterampilan)
     {
         $request->validate([
-            'nama_kategori' => 'required|max:255',
-        ]);
+    'nama_kategori' =>
+        'required|max:255|unique:kategori_keterampilans,nama_kategori,' .
+        $kategori_keterampilan->id,
+],[
+    'nama_kategori.required' =>
+        'Nama kategori wajib diisi.',
+
+    'nama_kategori.unique' =>
+        'Kategori sudah ada.',
+
+    'nama_kategori.max' =>
+        'Nama kategori maksimal 255 karakter.',
+]);
 
         $kategori_keterampilan->update($request->only('nama_kategori'));
 
@@ -59,11 +80,36 @@ class KategoriKeterampilanController extends Controller
             ->with('success', 'Kategori keterampilan berhasil diperbarui.');
     }
 
-    public function destroy(KategoriKeterampilan $kategori_keterampilan)
-    {
-        $kategori_keterampilan->delete();
+    public function destroy(
+    KategoriKeterampilan $kategori_keterampilan
+)
+{
+    if(
+        $kategori_keterampilan
+            ->keterampilans()
+            ->count() > 0
+    ){
 
-        return redirect()->route('admin.kategori-keterampilan.index')
-            ->with('success', 'Kategori keterampilan berhasil dihapus.');
+        return redirect()
+            ->route(
+                'admin.kategori-keterampilan.index'
+            )
+            ->with(
+                'error',
+                'Kategori sedang digunakan dan tidak dapat dihapus.'
+            );
+
     }
+
+    $kategori_keterampilan->delete();
+
+    return redirect()
+        ->route(
+            'admin.kategori-keterampilan.index'
+        )
+        ->with(
+            'hapus',
+            'Kategori berhasil dihapus.'
+        );
+}
 }
